third_party/WebKit/LayoutTests/platform/win/external/wpt/web-animations/timing-model/timelines/document-timelines-expected.txt

This is a testharness.js-based test.
FAIL Document timelines report current time relative to navigationStart assert_equals: The current time matches requestAnimationFrame time expected 19.31999999942491 but got 26.068666666105855
Harness: the test ran to completion.
